% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMMDIT.R
\name{addKnownHaplotypes}
\alias{addKnownHaplotypes}
\title{reads empop; adds haplotypes to database
This function takes an empop file containing at least four (mandatory) columns
(refer to https://empop.online/downloads for the emp file format)
and returns a tibble with two columns - SampleID and Variant}
\usage{
addKnownHaplotypes(db, empopFile)
}
\arguments{
\item{db}{database handle to MMDIT}

\item{empopFile}{a file of empop sequences...}
}
\description{
reads empop; adds haplotypes to database
This function takes an empop file containing at least four (mandatory) columns
(refer to https://empop.online/downloads for the emp file format)
and returns a tibble with two columns - SampleID and Variant
}
