% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countTraversalsMissedNodes}
\alias{countTraversalsMissedNodes}
\title{For each set of traversals, will count the number of missed nodes.}
\usage{
countTraversalsMissedNodes(sgrap, toExS, travISet)
}
\arguments{
\item{sgrap}{The graph.}

\item{toExS}{The graph traversals.}

\item{travISet}{The sets of traversals.}
}
\value{
The number of nodes missed by each set of traversals.
}
\description{
For each set of traversals, will count the number of missed nodes.
}
