% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utpalFunctions.R
\name{createPlafPan}
\alias{createPlafPan}
\title{A helper function to runDeploid()that creates panel and plaf dataframes}
\usage{
createPlafPan(EmpopLong)
}
\arguments{
\item{EmpopLong}{dataframe with information on sample ID, position and allele for a set of individuals}
}
\description{
This function takes a long dataframe (containing differences in coding compared to the rCRS,
in a select set of individuals) and returns panel and plaf dataframes
}
\examples{
\dontrun{
# createPlafPan(EmpopLong)
}

}
