% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utpalFunctions.R
\name{Empop2variant}
\alias{Empop2variant}
\title{reads empop
This function takes an empop file containing at least four (mandatory) columns
(refer to https://empop.online/downloads for the emp file format)
and returns a tibble with two columns - SampleID and Variant}
\usage{
Empop2variant(EMPOP, s = 1)
}
\arguments{
\item{EMPOP}{An EMPOP file (tab seperated)}

\item{s}{a numeric argument that tells the function how many rows of data to skip while reading in EMPOP file
If no value is provided, the first line is ommited by default}
}
\description{
reads empop
This function takes an empop file containing at least four (mandatory) columns
(refer to https://empop.online/downloads for the emp file format)
and returns a tibble with two columns - SampleID and Variant
}
\examples{
#   Empop2variant("EMPOP.emp")
#   Empop2variant("EMPOP.emp", s = 3)
}
