% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{enumerateTraversalPossibilities}
\alias{enumerateTraversalPossibilities}
\title{Enumerate the possible sets of traversals.}
\usage{
enumerateTraversalPossibilities(toExS)
}
\arguments{
\item{toExS}{The graph traversals.}
}
\value{
The possible collective traversal indices.
}
\description{
Enumerate the possible sets of traversals.
}
