% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semicontinuousMixtureInterpretation.R
\name{estimateLog10LikelihoodTheta}
\alias{estimateLog10LikelihoodTheta}
\title{Theta-corrected likelihood computation}
\usage{
estimateLog10LikelihoodTheta(
  allelesObserved,
  allelesThatExplain,
  theta,
  allelesInPopulation,
  populationCounts = NULL
)
}
\arguments{
\item{allelesObserved}{vector of alleles observed (empirically)  (not database, may be empty)}

\item{allelesThatExplain}{vector of alleles that explain the mixture (unknown haplotypes only)}

\item{theta}{theta-correction (Fst, taken from fst_buckleton)}

\item{allelesInPopulation}{a vector of alleles sampled from the population (also from database, but may in practice come from a different sub-population)}

\item{populationCounts}{a vector of integer weights (parallel to allelesInPopulation)}
}
\description{
This computes the theta corrected likelihood of a set of alleles hypothesized (unknowns)
given a set of observed haplotypes (allelesObserved) as well as a sample of alleles from the relevant population
and an estimate of co-ancestry.
}
\details{
This estimator is equation 3 in Ge et al. 2010 (doi.org/10.1016/j.legalmed.2010.02.003)

at a high level this computes the likelihood of a set of alleles hypothesized to explain some mixture
Some are there by the hypothesis (a victim -> contributing allelesObserved)
and others come from a database (unknowns)
as well as a sample of alleles from the relevant population (allelesInPopulation)

This estimates the allele frequencies (point estimate from known + unknown + population combined)
and weighs the likelihood of the alleles observed based on whether or not the unknowns and the knowns are identical
(x-vector).
A theta correction is also employed, where theta is an estimate of Fst (fst_buckleton)

It takes in a set of allelesObserved to be in the mixture (according to the hypothesis)
}
