% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMMDIT.R
\name{estimateLogLikelihood}
\alias{estimateLogLikelihood}
\title{Multi-population log-likelihood estimation}
\usage{
estimateLogLikelihood(
  allelesObserved,
  allelesThatExplain,
  databaseAlleles,
  databasePopulations,
  theta = NULL,
  ...
)
}
\arguments{
\item{allelesObserved}{vector of alleles observed (empirically) (not database, may be empty)}

\item{allelesThatExplain}{vector of alleles that explain the mixture}

\item{databaseAlleles}{a vector of alleles sampled from the population (also from database, but may in practice come from a different sub-population)}

\item{databasePopulations}{a vector of population identifiers (paired with databaseAlleles; also from database)}

\item{theta}{theta-correction (Fst, taken from fst_buckleton)}

\item{...}{(passed to estimateLog10LikelihoodNaive in the case that theta is null)}
}
\description{
This computes the P(evidence|Hypothesis) for a series of alleles
(allelesObserved, taken from known haplotypes and allelesThatExplain, populated from database)
which together explain the evidence.
}
\details{
The likelihood is estimated in multiple populations. To do so, provide parallel vectors of alleles (databaseAlleles)
and the population affiliations (databasePopulations). Each are treated as categorical variables (case sensitive,
either the same or different).

If a theta value is provided, the theta-corrected likelihood is estimated
(see: estimateLog10LikelihoodTheta)
otherwise the naive estimator is used.
}
