% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semicontinuousMixtureInterpretation.R
\name{estimateTheta}
\alias{estimateTheta}
\title{Estimates Fst and quantiles of Fst}
\usage{
estimateTheta(
  alleles,
  populations,
  quantile = 0.95,
  nJack = 0,
  approximate = FALSE
)
}
\arguments{
\item{alleles}{strings ; alleles (haplotypes);  1 row per chromosome sampled}

\item{populations}{strings ; same length as alleles (parallel array); population label associated with said haplotype}

\item{quantile}{; [0,1] ; what quantile in the Fst distribution would you like?}

\item{nJack}{; non-negative integer ; number of leave-one-out jackknifes taken}

\item{approximate}{; boolean ; whether or not allele frequencies (approximate=TRUE) or site hetero/homozygosities (FALSE) should be used}
}
\description{
This function computes Buckleton's estimator of Fst
(theta)
and it computes an upper-bound. It's a friendly
wrapper for fst_buckleton
}
