% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expandTraversalEditSequence}
\alias{expandTraversalEditSequence}
\title{Get the edit sequence a traversal represents.}
\usage{
expandTraversalEditSequence(sgrap, toEx, travInd)
}
\arguments{
\item{sgrap}{The graph.}

\item{toEx}{The traversal set in question.}

\item{travInd}{The index of the traversal set.}
}
\value{
The edit script to go from the query sequence to the reference graph. Three parallel vectors: the first is the location in the (current state of the) query, the second is the operation type (1 for substitution, 2 for insertion, 3 for deletion), and the third is the character (is garbage for deletion).
}
\description{
Get the edit sequence a traversal represents.
}
