% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expandTraversalGraphDifferences}
\alias{expandTraversalGraphDifferences}
\title{Get the variations from the reference that a traversal through the graph represents.}
\usage{
expandTraversalGraphDifferences(refRS, sgrap, toEx, travInd)
}
\arguments{
\item{refRS}{The reference sequence.}

\item{sgrap}{The graph.}

\item{toEx}{The traversal set in question.}

\item{travInd}{The index of the traversal set.}
}
\value{
The variants for that traversal through the graph. Three parallel vectors: the first is the location in the reference, the second is the operation type (1 for substitution, 2 for insertion, 3 for deletion), and the third is the relevant character/sequence (is garbage for deletion).
}
\description{
Get the variations from the reference that a traversal through the graph represents.
}
