% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMMDIT.R
\name{getAllDistances}
\alias{getAllDistances}
\title{generates amplicon sequence data from difference encodings}
\usage{
getAllDistances(amps, seqs, stops, ignoreIndels = FALSE)
}
\arguments{
\item{amps}{the output from getAmps}

\item{seqs}{query sequences}

\item{stops}{the stop coordinate of said sequences}

\item{ignoreIndels}{boolean; must match amps}
}
\description{
generates amplicon sequence data from difference encodings
}
