% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMMDIT.R
\name{getAmpCoordinates}
\alias{getAmpCoordinates}
\title{Get amplicon coordinates}
\usage{
getAmpCoordinates(db, kit = default_kit)
}
\arguments{
\item{db}{the database connection}

\item{kit}{the kit name (must be a valid KIT in the amps table)}
}
\value{
returns data frame with 4 columns: the amp ID, the start coordinate (0-based), the stop coordinate (1-based)
}
\description{
Generates the amplicon coordinates
for some kit
For use with MMDIT only!
}
\examples{
\dontrun{
amps <- getAmpCoordinates(db, kit="PrecisionID")
}
}
