% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMMDIT.R
\name{getAmps}
\alias{getAmps}
\title{makes amps}
\usage{
getAmps(db, pop = "\%", ignoreIndels = FALSE, kit = default_kit, blk = c())
}
\arguments{
\item{db}{the database handle}

\item{pop}{the population requested (defaults to all). vectorized populations okay}

\item{ignoreIndels}{strips out indel events (default TRUE)}

\item{kit}{the amplicon kit}

\item{blk}{blacklist of sites to filter out!}
}
\description{
generates amplicon sequence data from difference encodings
}
