% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utpalFunctions.R
\name{getdEploidHaps}
\alias{getdEploidHaps}
\title{Get estimated haplotypes based on the dEploid output}
\usage{
getdEploidHaps(dEploid.run, AR)
}
\arguments{
\item{dEploid.run}{dEploid output}

\item{AR}{Table of alternate and reference allele counts (tab-delimited plain text file, Empop2AltRef object)}
}
\description{
This function takes the output of dEploid and an Altref file and returns estimated haplotypes for the major
and minor contributors and their genomic positions
}
