% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMMDIT.R
\name{getMitoGenomes}
\alias{getMitoGenomes}
\title{gets full mitogenomes}
\usage{
getMitoGenomes(db, pop = "\%", ignoreIndels = FALSE, blk = c())
}
\arguments{
\item{db}{the database handle}

\item{pop}{the population requested (defaults to all). vectorized populations okay}

\item{ignoreIndels}{strips out indel events (default FALSE)}

\item{blk}{exclusion list of sites to filter out!}
}
\description{
generates amplicon sequence data from difference encodings
}
