% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utpalFunctions.R
\name{getMixProps}
\alias{getMixProps}
\title{Get proportion of mixtures from dEploid output}
\usage{
getMixProps(dEploid.run)
}
\arguments{
\item{dEploid.run}{dEploid object}
}
\description{
This function takes the output of dEploid and returns mixture proportions for major and minor contributors
}
\examples{
\dontrun{
getMixProps(dEploid.run)
}
}
