% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMMDIT.R
\name{getMtgenomeLength}
\alias{getMtgenomeLength}
\title{Length of mito-genome (reference)}
\usage{
getMtgenomeLength(db)
}
\arguments{
\item{db}{the database handle}
}
\description{
This is the non-circularized (concatenated)
mtgenome length (16569 unless some real funny-business is happening)
}
