% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMMDIT.R
\name{getMtgenomeSequence}
\alias{getMtgenomeSequence}
\title{Mito-genome (reference) sequence}
\usage{
getMtgenomeSequence(db)
}
\arguments{
\item{db}{the database handle}
}
\description{
Returns the self-concatenated mitochondrial reference genome
Self-concatenation allows us to handle the circular alignment problem
(eg, reads, amplicons that span the canonical "start" of the mt-genome)
and coordinates from a circular alignment
}
