% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMMDIT.R
\name{getNN}
\alias{getNN}
\title{finds nearest neighbors}
\usage{
getNN(alldist, knn = 20, maxdist = -1, useIdentityFunction = FALSE)
}
\arguments{
\item{alldist}{the output from getAllDistances}

\item{knn}{the number of nearest neighbors to find (ties broken arbitrarily)}

\item{maxdist}{if >-1, returns all neighbors within distance <= maxdist (ignores knn argument)}

\item{useIdentityFunction}{counts the number of amplicons that differ, not the sum of the distances}
}
\description{
finds nearest neighbors
}
