% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMMDIT.R
\name{getPops}
\alias{getPops}
\title{gets population labels}
\usage{
getPops(db)
}
\arguments{
\item{db}{database handle}
}
\description{
This takes in the database handle
and returns a the unique populations
}
