% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMMDIT.R
\name{getRefAmpSeqs}
\alias{getRefAmpSeqs}
\title{Amplicon sequences of reference}
\usage{
getRefAmpSeqs(db, kit = default_kit)
}
\arguments{
\item{db}{the database connection}

\item{kit}{the kit name (must be a valid KIT in the amps table)}
}
\value{
returns data frame with 4 columns: the amp ID, the start coordinate (0-based), the stop coordinate (1-based), and the Seq(uence) of the rcrs for the amp
}
\description{
Generates the amplicon sequences
based on the reference sequence.
For use with MMDIT only!
}
\examples{
# amps <- getRefAmpSeqs(db, kit="PrecisionID")

}
