% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getTraversalEditDistances}
\alias{getTraversalEditDistances}
\title{Will return the edit distances for the possible alignments to a graph.}
\usage{
getTraversalEditDistances(toEx)
}
\arguments{
\item{toEx}{The graph traversals.}
}
\value{
The edit distance of each traversal: length zero means no valid traversals were possible.
}
\description{
Will return the edit distances for the possible alignments to a graph.
}
