% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getTraversalHitNodes}
\alias{getTraversalHitNodes}
\title{Will return the nodes in the graph that the selected traversal hit..}
\usage{
getTraversalHitNodes(sgrap, toEx, travInd)
}
\arguments{
\item{sgrap}{The graph.}

\item{toEx}{The graph traversals.}

\item{travInd}{The index of the traversal to examine (one-indexed).}
}
\value{
The graph nodes hit by the traversal.
}
\description{
Will return the nodes in the graph that the selected traversal hit..
}
