% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getTraversalsEditDistances}
\alias{getTraversalsEditDistances}
\title{Vectorized version of getTraversalEditDistances.}
\usage{
getTraversalsEditDistances(toExS)
}
\arguments{
\item{toExS}{The graph traversals.}
}
\value{
The edit distance of each traversal: length zero means no valid traversals were possible.
}
\description{
Vectorized version of getTraversalEditDistances.
}
