% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getTraversalsHitNodes}
\alias{getTraversalsHitNodes}
\title{Will return the nodes in the graph that the selected traversals collectively hit.}
\usage{
getTraversalsHitNodes(sgrap, toExS, travInds)
}
\arguments{
\item{sgrap}{The graph.}

\item{toExS}{The graph traversals.}

\item{travInds}{The indices for each traversal.}
}
\value{
The graph nodes hit by the traversals.
}
\description{
Will return the nodes in the graph that the selected traversals collectively hit.
}
