% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utpalFunctions.R
\name{harmonizeAR}
\alias{harmonizeAR}
\title{A helper function to runDeploid()that harmonizes the AltRef}
\usage{
harmonizeAR(AltRef, p)
}
\arguments{
\item{AltRef}{a dataframe of reference and alternate allele counts (modified createAltRef object)}

\item{p}{vector of sites from a plaf/pan dataframe (createPlafPan object)}
}
\description{
This function takes an AltRef (modified createAltRef object),
and a vector p of sites from a plaf/pan (createPlafPan object)
and matches or "harmonizes" the number of sites between both dataframes
}
\examples{
\dontrun{
# harmonize(AltRef,p)
}

}
