% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utpalFunctions.R
\name{harmonizePa}
\alias{harmonizePa}
\title{A helper function to runDeploid()that harmonizes the panel}
\usage{
harmonizePa(panel, q)
}
\arguments{
\item{panel}{a dataframe representing a reference panel for a set of individuals (createPlafPan object)}

\item{q}{vector of sites from an AltRef dataframe with reference and alternate allele counts (modified createAltRef object)}
}
\description{
This function takes a panel (createPlafPan object),
and a vector q of sites from am AltRef (modified createAltRef object)
and matches or "harmonizes" the number of sites between both dataframes
}
\examples{
# harmonize(plaf,q)
}
