% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utpalFunctions.R
\name{harmonizePl}
\alias{harmonizePl}
\title{A helper function to runDeploid()that harmonizes the plaf}
\usage{
harmonizePl(Plaf, q)
}
\arguments{
\item{Plaf}{a dataframe of population level allele frequencies (createPlanPlaf object)}

\item{q}{vector of sites from an AltRef dataframe with reference and alternate allele counts (modified createAltRef object)}
}
\description{
This function takes a plaf (createPlafPan object),
and a vector q of sites from am AltRef (modified createAltRef object)
and matches or "harmonizes" the number of sites between both dataframes
}
\examples{
\dontrun{
# harmonize(plaf,q)
}

}
