% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utpalFunctions.R
\name{Hmtdb2Empop}
\alias{Hmtdb2Empop}
\title{Makes a giant empop file...}
\usage{
Hmtdb2Empop(LUT, Pa)
}
\arguments{
\item{LUT}{A lookup-table file (tab seperated)}

\item{Pa}{The path to folder where EMPOP files are stored (charachter string)}
}
\description{
This function takes a lookup-table file and a path to EMPOP files
and returns a bigger EMPOP file (i.e. all EMPOP files in the path bound together)
Note: This function is not for general use. It was constructed specifically
to construct haplotypes from HmtDB (https://www.hmtdb.uniba.it/) using custom up-stream processing.
}
\examples{

# Hmtdb2Empop(LUT, Pa = R.home())


}
