% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{listGraphNodes}
\alias{listGraphNodes}
\title{Lists the nodes in a sequence graph.}
\usage{
listGraphNodes(sgrap)
}
\arguments{
\item{sgrap}{The graph.}
}
\value{
The graph node indices.
}
\description{
Lists the nodes in a sequence graph.
}
