% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMMDIT.R
\name{loadAmpData}
\alias{loadAmpData}
\title{this loads in amplicon data (i.e., the coordinates of the amps)
into the MITO db.
If the alignments were to a modified reference (duplicating bits to accommodate circular alignment)
this routine converts those back to the original units.
thus the last amplicon may "wrap around" (eg., 16541-80 is a valid locus)}
\usage{
loadAmpData(db, ampFile, kitName, sep = "\\t", append = TRUE)
}
\arguments{
\item{db}{(the database connection)}

\item{sep}{(\"\\t\" for a tab-separated file. Anything else is a little wacky... no longer a bed file)}

\item{append}{(whether or not the amplicons are APPENDED to the list of current amplicons; set to FALSE to overwrite)}

\item{ampfile}{(a BED file of amplicons; columns 2 and 3 are used only)}

\item{kitname}{(the NAME of the kit that this corresponds to (e.g., PrecisionID) )}
}
\value{
the database handle (not necessary to keep this)
}
\description{
this loads in amplicon data (i.e., the coordinates of the amps)
into the MITO db.
If the alignments were to a modified reference (duplicating bits to accommodate circular alignment)
this routine converts those back to the original units.
thus the last amplicon may "wrap around" (eg., 16541-80 is a valid locus)
}
\examples{
# One way to add another kit:

# loadAmpData(db, "data/Mybedfile.bed", kit="SomeOtherKit", append=TRUE)


}
