% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMMDIT.R
\name{loadMMDIT}
\alias{loadMMDIT}
\title{Loads MMDIT}
\usage{
loadMMDIT(path = dataDir, dbFilename = dbName)
}
\arguments{
\item{path}{a directory (e.g., data)}

\item{dbFilename}{(e.g., mmdit.sqlite3)}
}
\description{
This *loads* the MMDIT database
path can either be the full path of the sqlite3 file
 (dbFilename must be "")
OR
path can be the directory
and the database name can be supplied separately
}
