% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMMDIT.R
\name{loadReferenceGenome}
\alias{loadReferenceGenome}
\title{Use this to add the rCRS sequence to MMDIT
It adds the record to the table mtgenome
this table has *1* row in it. The reference genome.
If you call this function multiple times, the last time
will be the record that "sticks"}
\usage{
loadReferenceGenome(db, fastaFile)
}
\arguments{
\item{db}{(database handle)}

\item{fastafile}{(must be both the name of the file and the path to open it. This is passed directly to seqinr::read.fasta}
}
\description{
Technically any genome would work here, BUT
the difference encodings MUST be with respect to this reference genome!!
i.e., in all likelihood this should *ONLY* be the rCRS sequence.
}
\examples{

# loadReferenceGenome(db, paste(dataDir, refGenome, sep="/") )

}
