% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMMDIT.R
\name{makeDB}
\alias{makeDB}
\title{Creates (an empty) MMDIT database}
\usage{
makeDB(path, dbFilename, overwrite = FALSE)
}
\arguments{
\item{overwrite}{(if there's already a file with that name, should I overwrite it?)
and creates a database with the requisite tables in it
if overwrite is FALSE, then it won't overwrite an existing database
in that path.
otherwise, it will clobber what was there (use overwrite=TRUE with caution!!)}

\item{path;}{a directory (e.g., data)}

\item{a}{filename (e.g., mmdit.sqlite3)}
}
\value{
the handle to the database (throughout all other documentation, this is db)
}
\description{
This *creates* an EMPTY MMDIT database
}
\examples{
# Make a new MMDIT database...

# db <- makeDB(dataDir, dbName, overwrite=TRUE )
# add the reference genome (can be only 1)
# loadReferenceGenome(db, paste(dataDir, refGenome, sep="/") )

}
