% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMMDIT.R
\name{makeDeploidSeqGraph}
\alias{makeDeploidSeqGraph}
\title{seqgraphs!}
\usage{
makeDeploidSeqGraph(db, positions, alleles, types, ignoreIndels = TRUE)
}
\arguments{
\item{db}{the database handle}

\item{positions}{positions of alleles in mixture (1-based)}

\item{alleles}{allele calls in mixture (1-based)}

\item{types}{type of variation; X == mismatch, I == insertion, D == deletion.}

\item{ignoreIndels}{Must be set to true. Correctness is only guaranteed if indels are ignored!}
}
\description{
generates a sequence graph for NN search
the input is a mitochondrial mixture (in "empop long" format)
the output is a seqgraph (from Haplotypical)
}
