% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMMDIT.R
\name{makeHaplotypeTable}
\alias{makeHaplotypeTable}
\title{Creates temp table of haplotypes}
\usage{
makeHaplotypeTable(
  db,
  sampleid,
  stopCoord,
  seq,
  tableName = "haplotypes",
  vtableName = "sequences"
)
}
\arguments{
\item{db}{(database handle)}

\item{sampleid}{(individual sample IDs from MMDIT)}

\item{tableName}{(the name of the table to be made; default="haplotypes")}

\item{vtableName}{(the name of the virtual spellfix table to be made; default="sequences")}

\item{stop}{(stop coordinate of amplicon)}

\item{sequence}{(the DNA sequence of this individual for this amplicon)}
}
\description{
this creates a temporary table called haplotypes
The haplotypes table converts the difference encodings into strings (after applying the mask)
}
