% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMMDIT.R
\name{makeSampleTable}
\alias{makeSampleTable}
\title{Creates temp table of *sample* haplotypes}
\usage{
makeSampleTable(db, stopCoord, seq, tableName = "sampleamps")
}
\arguments{
\item{db}{(database handle)}

\item{stopCoord}{(stop coordinate of amplicon)}

\item{seq}{(the DNA sequence of this individual for this amplicon)}

\item{tableName}{(the name of the table to be made)}
}
\description{
this creates a temporary table called refamps
The haplotypes table converts the difference encodings into strings (after applying the mask)
}
