% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{makeVariantGraph}
\alias{makeVariantGraph}
\title{Given a description of the variants in a mixture, will generate a graph for use with later analysis.}
\usage{
makeVariantGraph(refRS, varS, varE, varSeq)
}
\arguments{
\item{refRS}{The reference sequence.}

\item{varS}{The starting points of the variant regions.}

\item{varE}{The ending points of the variant regions.}

\item{varSeq}{The sequences for each variant.}
}
\value{
An opaque object containing the graph.
}
\description{
Given a description of the variants in a mixture, will generate a graph for use with later analysis.
}
