% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMMDIT.R
\name{naiveGetAllNN}
\alias{naiveGetAllNN}
\title{solves the All 1NN}
\usage{
naiveGetAllNN(amps, ignoreIndels = FALSE)
}
\arguments{
\item{amps}{the output from getAmps}

\item{ignoreIndels}{(whether or not indels were used to constitute the strings...)}
}
\description{
(much) more efficient solutions exist (cover trees)
O(n log n) for n individuals and a constant genome length
the solution provided is O(n^2) by the same count.
}
\details{
This looks at each individual in amps
and finds the 2NN of that individual and the associated distances
and the selects the 2nd individual. in theory this could give you
the same individual as you queried (ties are broken arbitrarily).
If you care about that... write your own :)
I just want the distances!
}
