% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semicontinuousMixtureInterpretation.R
\name{preprocessMitoGenomes}
\alias{preprocessMitoGenomes}
\title{Preprocess mitochondrial genomes}
\usage{
preprocessMitoGenomes(genomes)
}
\arguments{
\item{genomes}{a data frame from MMDIT::getMitoGenomes}
}
\description{
This takes the output of: getMitoGenomes
and indexes it. In particular, it extracts the unique haplotypes,
adds the haplotype ID to the genomes themselves, and it returns a list of
exactly 2 items.
The first is a dataframe with the genomes data frame
with an n column added (# of times the haplotype was seen) as well as a SeqID (unique ID)
the second is a dataframe with the haplotype sequence, n (# of times seen in database, total),
and the same SeqID
}
