% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utpalFunctions.R
\name{rundEploid}
\alias{rundEploid}
\title{Wrapper function for dEploid}
\usage{
rundEploid(AR, long, NumMCMC = 800, exportPostProb = TRUE, recomb = 0, k = 5)
}
\arguments{
\item{AR}{Table of alternate and reference allele counts (Empop2AltRef object)}

\item{long}{Table with information on sample ID, position, allele and type of mutation}

\item{NumMCMC}{A numeric argument that tells the function the number of MCMC samples (default value 800)}

\item{exportPostProb}{Save the posterior probabilities of the final iteration of all individuals}

\item{recomb}{numeric argument that gives the function the constant recombination probability (default value of 0.0)}

\item{k}{A numeric argument that tells the function the number of individuals in the mixture (default value 5, maximum of 5)}
}
\description{
This function takes information on alternate and reference allele counts,
number of individuals and the recombination rate and runs
the dEploid function on them. It retuns the dEploid output
with includes estimated haplotypes, mixture proprotions and other phasing statistics.
Additionaly this function also returns a list of nucleotide positions that violate the infinte sites model.
}
