% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utpalFunctions.R
\name{Snp2variant}
\alias{Snp2variant}
\title{Get variants from SNP data}
\usage{
Snp2variant(Pos, Allele, Type)
}
\arguments{
\item{Pos}{A vector of genomic positions for the SNPs (type numeric)}

\item{Allele}{A vector of nucleotide bases present in the SNPs (character strings). Deletions are represented by ""}

\item{Type}{A vector of the type of mutation represented by the SNP. Possible values include "Substitution", "Insertion"
and "Deletion" (charachter strings)}
}
\description{
This function takes three vectors Position, Allele and Type from the SNPs
and returns a vector of variants (e.g. "73G", "73+G", 73-" )
}
\examples{
Snp2variant("73", "G", "Substitution")
Snp2variant(c("73", "95", "146"), c("G", "", "C"), c("Substitution", "Deletion", "Insertion"))
}
