% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{str2int}
\alias{str2int}
\title{strings to ints}
\usage{
str2int(s, out)
}
\arguments{
\item{s}{(vector of strings)}

\item{out}{(vector of integers; this is filled out by the method; same length as s)}
}
\value{
number of distinct/unique strings
}
\description{
This is a Rcpp implementation of R's
as.integer(as.factor(strings)) - 1
}
\details{
ie, it converts strings into integers (0-based)
the conversion maintains the natural ordering
(the first string is 0, the next new string i 1, ... )

returns number of unique elements
}
