% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMMDIT.R
\name{testEditdist}
\alias{testEditdist}
\title{Trivial testing to see if the editdist3 function works...
Takes no arguments; returns a data frame (if editdist3 works!)...
If this *doesn't* work you need to make sure you're
using the development version of rsqlite as downloaded from my github at:}
\usage{
testEditdist()
}
\description{
devtools::install_github("Ahhgust/RSQLite", ref="devel", upgrade=TRUE)
}
