% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semicontinuousMixtureInterpretation.R
\name{threepersonMix}
\alias{threepersonMix}
\title{3 person mixture simulations}
\usage{
threepersonMix(
  db,
  pops = c("EU"),
  seed = 1,
  nMixes = 1000,
  ignoreIndels = FALSE
)
}
\arguments{
\item{db}{database handle}

\item{pops}{population groups to use.}

\item{seed}{sets the seed in the random number generator}

\item{nMixes}{the number of simulations to do}

\item{ignoreIndels}{when TRUE, insertion/deletions are omitted}
}
\description{
This simulates 3-person mixtures from the populations (pops) specified
and it creates a data frame with summary statistics on said mixtures.
RMNE type statistics are returned, as well as RMP/LR statistics (or the basis thereof)
}
