% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{traverseSequenceGraph}
\alias{traverseSequenceGraph}
\title{Will return all alignments of a sequence to a graph (within some maximum edit distance).}
\usage{
traverseSequenceGraph(sgrap, query, maxEdit)
}
\arguments{
\item{sgrap}{The graph.}

\item{query}{The sequence.}

\item{maxEdit}{The maximum edit distance.}
}
\value{
An opaque object containing the graph traversals.
}
\description{
Will return all alignments of a sequence to a graph (within some maximum edit distance).
}
