% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{traverseSequencesGraph}
\alias{traverseSequencesGraph}
\title{Vectorized version of traverseSequenceGraph.}
\usage{
traverseSequencesGraph(sgrap, queries, maxEdit)
}
\arguments{
\item{sgrap}{The graph.}

\item{queries}{The sequences.}

\item{maxEdit}{The maximum edit distance.}
}
\value{
A list of opaque objects containing the graph traversals.
}
\description{
Vectorized version of traverseSequenceGraph.
}
