% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semicontinuousMixtureInterpretation.R
\name{twopersonMixOriginal}
\alias{twopersonMixOriginal}
\title{2 person mixture simulations}
\usage{
twopersonMixOriginal(db, pops = c("EU", "AM"), seed = 1, nMixes = 1000)
}
\arguments{
\item{db}{database handle}

\item{pops}{population groups to use.}

\item{seed}{sets the seed in the random number generator}

\item{nMixes}{the number of simulations to do}
}
\description{
This simulates 2-person mixtures from the populations (pops) specified
and it creates a data frame with summary statistics on said mixtures.
RMNE type statistics are returned, as well as RMP/LR statistics (or the basis thereof)
}
