% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utpalFunctions.R
\name{Variant2snp}
\alias{Variant2snp}
\title{Get SNPs from variant data}
\usage{
Variant2snp(Variant)
}
\arguments{
\item{Variant}{A vector of variants (character strings). See Snp2variant for description}
}
\description{
This function takes a vector of variants (e.g. "73G", "73+G", 73-" ) and returns a tibble
with three columns : Pos, Allele and Type that contain information on the genomic position,
nucleotide base and type of mutation (i.e. "Substitution", "Insertion" and "Deletion")resecptively for each variant
}
\examples{
# Variant2snp("73G")
# Variant2snp(c("73G", "95-", "146+C"))
}
