% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utpalFunctions.R
\name{write_mbop}
\alias{write_mbop}
\title{Makes an "empop" dataframe}
\usage{
write_mbop(SampleID, Variant)
}
\arguments{
\item{SampleID}{A vector of sample IDs (character strings)}

\item{Variant}{A vector of variants (character strings). See Snp2variant for description}
}
\description{
This function takes two vectors SampleID and Variant
and returns a tibble in the empop format.
Each row in the tibble will correspond to 1 individual (SampleID)
and their empop string (tab separated) will be the value in the second column
}
\examples{

# write_mbop("NA12871", c("73G", "95-"))
# write_mbop(c("NA12871", "NA12871", "NA12872"), c("73G", "95-", "73G"))

}
