% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMMDIT.R
\name{getSeqdiffs}
\alias{getSeqdiffs}
\title{get all seqdiffs for some population}
\usage{
getSeqdiffs(db, pop = "\%", ignoreIndels = FALSE, getPopulation = FALSE)
}
\arguments{
\item{db}{the database handle}

\item{pop}{the population requested (defaults to all). vectorized populations okay}

\item{ignoreIndels}{strips out indel events (default TRUE)}

\item{getPopulation}{gets the population label along with the sequence differences}
}
\description{
get all seqdiffs for some population
}
