% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMMDIT.R
\name{loadMMDIT}
\alias{loadMMDIT}
\title{Loads the database...}
\usage{
loadMMDIT(path = dataDir, dbFilename = dbName)
}
\arguments{
\item{path}{a directory (e.g., data)}

\item{dbFilename}{(e.g., mmdit.sqlite3)}
}
\description{
This *loads* the MMDIT database
The haplotypes table will be empty (but fillable)
}
