% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMMDIT.R
\name{makeHaplotypeTable}
\alias{makeHaplotypeTable}
\title{Creates temp table of haplotypes}
\usage{
makeHaplotypeTable(db)
}
\description{
this creates a temporary table called haplotypes
The haplotypes table converts the difference encodings into strings (after applying the mask)
}
\details{
#' @param db (database handle)
}
